requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.1
fonttools==4.58.1
greenlet==3.2.1
gunicorn
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
SQLAlchemy==2.0.40
toolz==1.0.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
flask-restx==1.3.0
